refactor(model): split Report.kt into separate files for better organization
- Moved classes into individual files: Report.kt, User.kt and Location.kt
- Added new data classes Farmer and Vet in their respective files
- Will need to change the imports in all the code in a later commit
