+++ AgriHealth-Alert-main/app/src/main/java/com/android/agrihealth/data/model/location/Location.kt
@@ -1,0 +1,3 @@
+package com.android.agrihealth.data.model.location
+
+data class Location(val latitude: Double, val longitude: Double, val name: String? = null)
