+++ AgriHealth-Alert-main/app/src/main/java/com/android/agrihealth/data/model/user/User.kt
@@ -1,0 +1,6 @@
+package com.android.agrihealth.data.model.user
+
+enum class UserRole {
+  FARMER,
+  VET
+}
